Refactor for display default content
